liblink: avoid overrunning symbol buffer in span7
Aligning DWORDs onto 8 byte boundaries can make a function take up
more space than the code expects and overrun the buffer.  This was
causing realloc() to abort with an "invalid pointer" error, I guess
because it was corrupting glibc's heap data.

This fixes by checking to see if the padding will cause the buffer to
overflow and resizing it again.

Change-Id: Ic4b2a1a478765d86812fd329762de8cd2ff7802d