cmd/7g: reactivate componentgen for TINTER
Update #80

7g and 9g do not use componentgen and so fall back to the generic behaviour in sgen which is broken, #80.

This PR reactivates parts of componentgen so sgen can use it for TINTER, which fixes #80 with respect to passing interface values to functions. All other fat types are still broken.

Arguably we could reactivate the other parts of componentgen but that would be masking a latent issue in generic sgen.

Anyway, this gets @mwhudson and I a little further, and I'll work on fixing sgen properly.