liblink, cmd/internal/obj: detect when the compilers wants $lauto from us
The compilers don't behave properly and expect liblink to figure out
a->name from an Addr with a->type=D_ADDR. In reality, a good citizen
would set a->name=D_AUTO when it wants the effective address of an
automatic. In fact, the compilers do this sometimes, but not always.

It is too hard to fix the compilers, and we don't want to touch non-arm64
code anyway, so we do for arm64 what other liblinks do. Unfortunately,
on arm64 this is complicated by the fact that ZR and SP are the same
register (R31). Another idiosyncrasy of liblink is that it represents
zero constants by setting a->reg=REGZERO, but that clashes with
REGSP. Fortunately, a->reg is set only when a->offset is zero, and
usually a->offset is non-zero when asking for effective stack address
(it makes no sense to ask for a zero offset, since then you can just
use the stack pointer itself with no constant), so we use a->offset as
a distinguishing marker.

In reality, there is no reason that $0 should use a different
encoding than $42, because aclass only cares about a->offset when
returning C_ZCON. If $0 would have a->reg=NREG (32), like other
constants, then we could use a->reg itself to distinguish D_NONE from
D_AUTO/D_PARAM. Unfortunately, the author's efforts to normalize the
encoding of constants have failed. If we use NREG for $0, the code is
missasembled in subtle ways. We abandon that effort for now, and we
continue with this hack.

Fixes #80