arm64, malloc: Support 64KB pages
On arm64 we can run with 64KB pages as well as 4KB pages.
Unfortunately, the memory allocator in go assumes pages are of size 4KB
and we run into problems such as:
    runtime: address space conflict: map(0x4000001000) = 0x3ff7e090000
    fatal error: runtime: address space conflict

This patch increases the _PageShift to 16 for arm64 and also beefs up
the _StackCacheSize to one page for arm64.

I have tested this patch on an AMD Seattle system running 64KB pages
and 4.0-rc5 kernel as well as an APM Mustang system running 4KB pages
and 3.19 kernel.

I don't know why this hasn't arisen in PowerPC, perhaps a lucky set
of addresses back from mmap? Perhaps this patch could be extended for
PowerPC too if need be?

Signed-off-by: Steve Capper <steve.capper@linaro.org>